+++ src/main/resources/templates/homePage.html
@@ -1,5 +1,5 @@
 <!DOCTYPE HTML>
-<html xmlns:th="https://www.thymeleaf.org">
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
 <head>
     <title>SurveyMonkey Home</title>
     <link rel="stylesheet" type="text/css" href="css/styles.css" />
@@ -12,8 +12,8 @@
     <h2>Please choose if you are a Surveyor or a Taker.</h2>
 
     <div class="buttons">
-        <button onclick="location.href='#'" type="button" >Surveyor</button>
-        <button onclick="location.href='#'" type="button">Taker</button>
+        <button onclick="location.href='/surveyorHome'" type="button" >Surveyor</button>
+        <button onclick="location.href='/takerHome'" type="button">Taker</button>
     </div>
 
 </div>
