+++ src/main/resources/templates/surveyorHome.html
@@ -1,0 +1,36 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Surveyor Home</title>
+    <link href="css/surveyorHome.css" rel="stylesheet" type="text/css"/>
+</head>
+<body>
+<div style="justify-content: center;text-align:center;">
+    <h1>Surveys</h1>
+    <h2>Created Surveys shown below</h2>
+    <div style="border: thin solid black; border-radius: 10px; padding: 20px; display: inline-block">
+        <h4>All Surveys</h4>
+        <table>
+            <th:block>
+                <tr>
+                    <td class="td"><h3>Survey 1</h3></td>
+                    <td class="td">     <a href="#">     Edit</a></td>
+                    <td class="td">
+                        <a href="surveyResults">     Results</a>
+                    </td>
+                </tr>
+            </th:block>
+
+        </table>
+    </div>
+
+    <div>
+        <h2><a class="h2" href="#">Create Survey</a></h2>
+    </div>
+</div>
+
+
+
+</body>
+</html>
